🚑 removes error tooltip which caused lock up (#716)
When we hovered over a tooltip on an error, it would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, which displayed a tooltip which would cause a re-render, stack overflow, lock up the app.

I fixed it by deleting it.  The tooltip wasn't providing much value.  It was showing the full path.  It rarely fit on the screen properly.